Refactor: 서비스 시작 종료 진입점 수정
